#52 Move existing basis-specific stub tests into dynamo0p3_basis_test.py
